Strict commit was a little .. ah .. too strict, oops :}
branch.unknowns() returns a generator, so we can't do if branch.unknowns(),
it's always True.

if len(list(branch.unknowns())) might look cleanest, but creating the list of
unknowns is not cheap for large numbers of unknown files.

Instead we use a for loop to raise an exception as soon as we get one unknown
back from the generator.

Add more tests to make sure this doesn't break again.