[patch] annotate --long (robey pointer)
This fixes a single TODO in 'bzr annotate' that has been bugging me:
the email field is truncated even with '--long'.

The fix is pretty tiny (as you can see).  It causes some generator
results to be collected instead of iterated, but this turns out to be
not a big deal, since the yield is very shallow and comes from a
collection one stack level deeper.