Better progress bars during tests
This now shows by default something like::

  [41/3989, 10 skipped in 0s] test_bundle.V08BundleTester.test_binary_bundle

The whole test name is generally visible, as are the number skipped or 
failed so far; if there are skips, failures or errors they are printed
out above the progress bar. 

Code for displaying test results in different ways was separated out
into cleaner subclasses, and the selftest builtin was cleaned up.

This introduces a new and (simpler) setup for progress bars through the UI
object.  They don't nest, under the assumption that we can really only convey
one level at a time.