Use repo.supports_rich_root() everywhere rather than repo._format.rich_root_data.
This is needed for the smart server work, because a RemoteRepository's _format
doesn't know whether or not the actual repository supports rich root data or
not.  But the RemoteRepository does (by asking the underlying repository it is
connected to).

This also removes some unnecessary isinstance(..., Repository), which also
hinders smart server work, because RemoteRepository doesn't inherit from
Repository, even though it implements the same interface.