Committing a debug thunk that was very helpful
find_seen_ancestors() should *not* be accessing the real_provider() as it
should only be looking at the part of the graph that has already been
accessed.

However, in some of our final checks, the list passed in included some of
the stuff we had only peaked at.

This drops the time for "bzr status" after a simple merge down from 5.2s down
to 1.76s. (Before this fix it was closer to 6-11s for my "improved" form).