Change _search_for_extra_common slightly.
The official statement is that you can stop searching common nodes
once all common tips are ancestors of *all* unique nodes.
We need to make sure that we stop the common searchers once
they find nodes that are ancestors, and then we can only
exit the common search loop when they have finished.

All tests pass, need to see if it is still 'efficient'.