Fix bug #214894. Fix RemoteRepository.get_parent_map() when server is <v1.2
When connecting to a server that is <v1.2, the get_parent_map() code falls back
to using get_revision_graph(). However, for nodes with no parents, get_revision_graph()
returns an empty tuple, while get_parent_map expects (NULL_REVISION,).
In get_parent_map() an empty tuple means the node is a ghost.

This caused problems pushing Branch5 branches, because it would check the revision history
and consider the first commit to be a ghost. Which meant it was excluded from the history.
And then len(revision_history) was always 1 shorter than the expected value.