Change the per-file log algorithm dramatically.
Now we use the merge_sort information to decide when to show
another entry. It seems this changes what is actually logged.
This is because sometimes a branch merges the trunk. And in that
case, it will be bringing in the changes to a file, which the
feature branch did not have before.
However, for people using 'bzr log file' I believe those are
uninteresting nodes.

Memory consumption is good (240MB, mostly from text_index._buffer_all),
speed is good (28s, also mostly in _buffer_all).

Changing the code to request one key at a time saves the
buffer all, but at the expense of lots of bisects. In some
cases it is faster and saves memory (23s, 88MB), but on
bzr.dev it is slower.