Change the logic to solve it in a different way.
Now autopack will always write out a single pack file. It uses
the same 'fitting' logic, to determine which packs are
sub-optimally sized and should be combined. However, when it
finally comes to packing them, it just puts them in the same file.
It is the same amount of I/O, it just leads to fewer final
pack files.

Some analysis should be done on the long-term effects of this.

As a side effect, it seems immune to bug #242510. The only way
to trigger it was to have a large pack built, and then wanting
a smaller pack. However, if you are genuinely building a pack,
then there is something to be combined with.