Part of bug #300289, stop requiring plain fulltexts for revisions.
The generic fetch code was always extracting the fulltext for revisions as part of
the fix for bug #261339, however insert_record_stream is currently perfectly
capable of upcasting to a fulltext on-the-fly.
So we restore the behavior of copying knit-ft-gz without extracting it,
and add another test that ensures if the source has a delta it gets expanded
to a fulltext in the target.