We can use 'random_id=True' when copying the streams. This is because the 'get_stream' code is responsible for ensuring the keys are truly non-overlapping, and we know we are creating a new pack file.
It might mean that we have some overlap with yet-another existing
pack file, but only if some other operation inserted it accidentally,
and that doesn't hurt anything. When we autopack or fetch, we will
skip one of those records anyway.

This saves quite a bit of time, since we don't have to look up
texts in the index we are writing. Mostly only in the case of
large projects where we have spilled some of the nodes to disk
already.