Add a per_repository_reference test with real stacked repos.
This adds a test that sets up a stacked repository with an extra revision
versus the stacked-on location, and then asserts you can branch from this.
I would think this would be tested somewhere, but it wasn't a 
per-repository-reference test, so when I broke the code, it didn't fail
an obvious test.
This now changes the code so that if there are no keys to be fetched for
the 'missing_keys', then it doesn't try to get a stream from that vf.
There wasn't any real reason to do so, and it avoids trying to get
a stream for chk_bytes when the vf doesn't exist.