Fix some failing tests for repositories and win32.
Old format repositories did not escape characters like ':', which is
not a valid filename on Windows. As such, we need to trap when 'tree.commit'
fails, and skip the test.

Also, updated the test to make a proper distiction between a 'rev_id' which
would be a string, and a 'rev_key' would would be a tuple.