Clean up the calls for '_create_inv_from_null' so they use the apis correctly.
In the end this shaves off as much as 2s (15.5s => 13.5s) for an initial commit of
a mysql tree. Some of that is potentially the InternalNode.map() fix, some of it is
not going through a regular Inventory and then back into an apply_delta loop, etc.

Stuff like InternalNode.map() has 2 node._current_size() calls, so that it can see
if the size changed so it knows to check for remap, etc, is wasted on an initial
build. And optimizing the 'build-from-scratch' is somewhat reasonable, since it is
the only time that we should be dealing with that many objects.