Adding a parent inventory cache, and then using add_inventory_by_delta.
We need both the text cache and the inventory cache, because we need the *text* cache
to be able to apply the mpdiff, and we need the *inventory* cache to be able to
compute the deltas.

With both, the time to insert 577 revisions drops to 1m23s.

This is the hit rate:
{'cached parent inv': 459,
 'cached parent text': 564,
 'missing parent inv': 98,
 'missing parent text': 149,
 'parents': 713}

so we have the parent text 564/713 times (pretty good), and the parent inventory
459/557 times, also quite good.