tsort.TopoSorter: performance improvements
_push_node and _pop_node are inlined and all the self._vars are made local
variables, like how it already was done for MergeSorter.  This results in a
~40% reduction of runtime for topo_sort.

Renamed _node_name_stack to pending_node_stack to make the relation with
pending_parents_stack more obvious.  Furthermore, some comments are added and
updated and a while loop is eliminated, to make the code structure more clear.