Notes for improvement on the OOo tree.
Better than 3x faster even when there is only a single pack file.

So sure, it would be nice to not load the whole ancestry, and that
is important to try to do. But it is also nice to be 3x faster in
those cases where we haven't figured out how to do less.