revno 4639 in bzr.dev introduced a bug in the conversion logic for 'IDS'.
This bug is present in the InterDifferingSerializer code path, which is
used for local conversions (e.g. bzr upgrade on local disk, or bzr
push/pull between different formats, on local disks).

This revision is present in the 2.0rc1 release.

The bug is simple: A variable used to inform the repository where a
delta should be applied from was aliased in a for loop, leaving its
value indeterminate.

In *some* cases this will cause later consistency checks to error, but
not in all cases, as we don't do a manifest check during conversion.

No commits as yet, but getting this out is a priority.