Implement tp_traverse, and add a soft dependency on meliae to test it.
Basically, gc.get_referents doesn't use tp_traverse if the object isn't being
tracked. Which is ok, but not what we want here.
Key isn't in GC because it is overkill for something that can never have
cyclical references.
So I implemented a get_referents in meliae that ignores IS_GC.