Add a _static_tuple_c.pxd file to define the C api to pyrex code.
I had to drop the notion of void*items[1] and then subtracting it from tp_basicsize
because Pyrex directly checks that sizeof(struct) == tp_basicsize. I understand
why it was doing so, but it was very hard to track down. The error during
import was basically suppressed, and then I just got segfaults because
the type wasn't actually available.

Anyway, I'm reasonably happy with using a .pxd file, because then I don't have
to worry about getting the invocation right in all callers, though I still
have to repeat the 'from x cimport a,b,c,d' which is a bit annoying.