Have a defined result for Tree.path_content_summary() for unversioned files.
WT.path_content_summary() was failing because it was trying to look up
the object in the working inventory, and that has no record for unversioned
content.
PreviewTree seems to be a bit indeterminant. Aaron claims it should also
return a 'file' and the info it knows about, but it currently returns 'missing'.
That is now marked as a 'knownFailure'.