Quick fix for glob tests.
It turns out that some filesystems (Vincent's build-bot) will return
the glob in strange orders. It is *allowed* by the filesystem not
to glob in sorted order, so we should sort afterwards.
This isn't strictly true if we had 'b/* a/*', then it shouldn't
return b after a, but for our testing we never try that, and it makes
life easier.