HTTPServer sets self.host from the socket address.
On a system with IPv6 enabled, it can try to connect to the ipv6 address
first, and then takes 1s to timeout, before trying the ipv4 address.
However, we know at 'bind' time that we are only binding to the ipv4 address.

The effect:
  time bzr selftest -s bt.test_sftp_trans -s bt.test_smart -s bt.test_smtp -s bt.test_http
changed from 500s down to 27s. May not have that effect everywhere,
but places that test against the http server have a rather large effect.