List alternatives for --authors option.
Unfortunately a normal RegistryOption doesn't seem to have a way of
including the help strings for its values in its own help text. 
RegistryOption.from_kwargs used to do so, but we want a real registry here,
with the actual implementation as its value, not only a string identifier.

As not listing the alternatives at all feels even worse than hardcoding them
into the message, and as dynamically generating the help text just in this
instance feels like an evil hack as well, we'll keep it this way for now,
until a suitable improvement to the RegistryOption class gets implemented.