Bug #375013, Basic support for filling in basis inventories
In short, when we go to commit to a stacked repository, we just check for the
presence of any required parent inventories. If they aren't there, we fetch
them from whatever fallbacks we have. We are able to use the Stream interface
which is what the stacking code uses today anyway.
The write-group handling is a bit ugly, but I'll bring that up in review.