Update the performance list in whats-new.
We generally have 3 bits that we improved

1) Make loading the whole inventory not be pathologically slow.
2) Make more cases use optimized iter_changes() results (ones that
   don't require loading the whole inventory)
3) Update the WT basis cache using a delta, rather than requiring it
   to be set from scratch.

(1) drops us from 3+minutes down to ~15s for each tree we have to
load.
(2) Mostly because we don't have to spend those 15s, we get down
to a couple seconds here and there
(3) This covers a few more cases that (2) didn't directly apply.
Stuff like 'pull' has to do iter_changes (which we made faster
in (2)) and then set_parent_trees (which was made faster in 3)