Use move instead of rename when adding packs to repository.
A rename might fail (e.g. on sftp) if the target file does exist. This is
because rename is guaranteed to be atomic.  The move operation, on the other
hand, does guarantee deletion of any existing target if possible, so it is
more appropriate here, where we are not dealing with locks but with content
files.