allow a bit more flexibilty in how we walk the searches.
Allow setting the depth to 0 to indicate we want to walk the whole graph.
Always use the limit form, there isn't much benefit to the other. It is
a bit faster, but when the graph is that small, the time doesn't really
matter.

Some data from various knob adjustments with lp:gcc-linaro:
 search_depth  time  total transferred
 0             332.6s  995kB
 1              69.1s 1408kB
 20             65.8s 1405kB
 1000           82.9s 1390kB
 10000         129.8s 1384kB

I was a bit surprised that '0' has such a large effect on bytes transferred,
however, you can see it comes at a very high runtime cost.
This is going on about 35ms ping which is pretty low, but it isn't the
loopback.

Similar numbers for lp:bzr:
 search_depth  time  total transferred
 0             43.2s 2558kB
 1             46.5s 2902kB
 10            42.6s 2601kB
 20            44.1s 2595kB
 100           43.3s 2574kB
 1000          45.4s 2575kB
 10000         44.8s 2575kB

The numbers are all pretty within noise margin. My guess is that the 6k depth
of bzr isn't sufficient to cause much O(N^2) headache. lp:gcc-linaro is in the
100k depth range, which is where it shows up dramatically. Even 10k revisions
starts to show up vs this latency. Note that these show up more dramatically
over the loopback.


bzr://localhost/bzr.dev:
 search_depth  time  total transferred
 0             16.3s 2561kB
 1             17.4s 2830kB
 10            16.4s 2603kB
 20            16.7s 2597kB
 100           16.1s 2575kB
 1000          17.3s 2577kB
 10000         17.3s 2577kB

Also note that local performance is still slightly off, because we do extra
work searching for redundant heads, and running the graph search a second
time. I've never had the assertions trigger, so I think it is safe to
remove the second graph search.

Somewhere in the 10-100 seems to be a pretty good 'sweet spot' between
latency and CPU time spent.

Another big graph that is more 'bushy' vs gcc/emacs very linear history:
lp:mysql
 search_depth  time  total transferred
 0             126.1 5031kB
 1             129.1 5595kB
 20            126.1 5143kB
 100           119.5 5091kB
 1000          136.0 5090kB
 10000         132.2 5090kB

Note that these numbers were only run 1 time over a network, so there any
variation could certainly be within noise margins. I did try not to run
other programs or load my network while doing it.

The key takeaway for bzr and mysql (IMO) is that as long as you aren't
at 1 or 0, the numbers are modestly insensitive. So there is variation,
but not enough to indicate a strong inflection point, etc. I think I
would be happy with 20 or 100 as a smart guess.