Just use max_size.
We get better results with a larger estimate, but our estimator isn't giving better
results, it is just that doing more-per-pass is pretty much always better.
The *big* win is that we have 9-round-trips for all of gcc's history rather
than 46 round trips.
For bzr, increasing the estimation just results is more data being sent. And
we start breaking the 64k desired size a bit too often.
Live with it for now.