Now that we have 2 possible locations where things exist, we must cleanup both.
test_kind_change was re-using the 'tree' location for each permutation of the test, and
using osutils.rmtree() to cleanup. However that won't clean up the RemoteRepository
location at all.
So switch to using tree.bzrdir.root_transport.delete_tree(), and also check if the
branch location needs to be cleaned up separately.