record_stream wasn't being iterated to completion.
This meant that we didn't always finalize a nested_progress_bar.
Which meant that our test suite thought we were at a depth >1.
Which meant that ProgressRecordingUIFactory was discarding
the update messages.
So wrap the progress bar in a try/finally. That way when the
generator exits, we get a proper cleanup.
Note that record_stream is generally driven by iterating the
stream and stepping it, so it isn't too surprising that
we weren't stepping that one final time to trigger generator
teardown.