Move read_bytes_from_socket into osutils like it is in bzr-2.5.
This makes sure we are catching the extra exceptions, etc.

The code is now essentially the same as 2.5 except for:
a) No class SmartClientAlreadyConnectedSocketMedium, since I don't want to bring in
the code that uses sockets to communicate to the subprocess.
b) No code for auto-disconnecting clients. That is more of a server-side change
and I only care to change the client side.
c) A fair number of differences that are just not related.