(bug #830497) Teach Dirstate._validate to check that id_index maps back to actual entries.
This makes it clear that set_root_id should not be updating the index, as update_minimal
is likely doing it correctly. (you don't want id_index to map to the unset root location,
only to the currently set locations.)

This makes the plain set_root_id test fail, which means we don't have to write
any new tests.