Backport the fix for bug #855155 to bzr-2.4
Dirstate.update_basis_by_delta had a logic flaw where if it got a delta that added files to
an otherwise empty directory, it would notice that the dirblock didn't exist,
and it would claim the delta was invalid, rather than noticing the parent
directory did exist, it just had an empty (thus nonexistent) block.
This never triggered in commit, because you always have the files in 'current'
which forces the block to exist. But things like uncommit can change the basis
without touching current.